Finish allowing distributions of any type

+++ counting-language-impl/src/main/java/com/github/mhdirkse/countlang/execution/DistributionItemItemCalculation.java
@@ -33,6 +33,6 @@ class DistributionItemItemCalculation extends ExpressionResultsCollector {
 
     @Override
     void processSubExpressionResults(List<Object> subExpressionResults, ExecutionContext context) {
-        builder.add((Integer) subExpressionResults.get(0));
+        builder.add(subExpressionResults.get(0));
     }
 }
